Fix the build
RedHat and/or the JBoss community still doesn't understand that a Maven
release artifact published to a public Maven repository should never be
deleted. That caused problems with some of the dependencies of RHQ
4.12.0 which were loaded from a "tech preview" repository and which have
been removed. 4.13.0 has the same problem, so switch to 4.12.0.JON330GA.
Unfortunately rhq-enterprise-server doesn't exist in that version, so
continue to use 4.12.0 for that artifact.